refactor(accounts): split accounts/tasks.py into a per-domain package (#607)
accounts/tasks.py was 995 lines — one feature away from the hard 1000-line CI
file-length gate. Split it into an accounts/tasks/ package with domain-focused
submodules (email, tokens, gdpr, verification_reminders, notifications, bans,
payouts). Pure refactor, no behavior change.

Celery task names are preserved exactly — this half needed more care than the
events split (#606) because six accounts beat-scheduled tasks rely on the
*default* module-path name `accounts.tasks.<func>`, referenced by string in
periodic-task migrations:
- All 19 bare `@shared_task` tasks are pinned to `name="accounts.tasks.<func>"`
  (their previous default name), so the registered name stays byte-identical
  across a rolling deploy. `process_referral_payouts` keeps its explicit
  `accounts.process_referral_payouts`. Verified at runtime: all 20 tasks
  register under their original names, zero submodule-path drift.

`accounts/tasks/__init__.py` re-exports every task plus the
`DATA_EXPORT_URL_EXPIRES_IN` constant, so all existing
`from accounts.tasks import <name>` import sites keep working untouched. It also
re-exports `mark_reminder_sent`: `common.tasks._execute_email_callback` resolves
that callback via `getattr(import_module("accounts.tasks"), "mark_reminder_sent")`
against its allowlist, so it must stay importable from the package (the
callers keep `callback_data["module"] = "accounts.tasks"`).

The only test changes retarget three module-global patches that follow their
task into a submodule:
- `accounts.tasks.stripe` / `._send_payout_statement_email` / `.send_email`
  → `accounts.tasks.payouts.*` (payout tests)
- `accounts.tasks.httpx` → `accounts.tasks.notifications.httpx` (admin-notif tests)
- `accounts.tasks.send_email` → `accounts.tasks.verification_reminders.send_email`
  (reminder tests)

Migration/beat sweep: all 28 scheduled task-name strings across every app's
migrations resolve against the live registry (forced full task discovery) —
UNRESOLVED: NONE. make check green (ruff, mypy --strict, migrations, i18n,
file-length); 210 affected account-task + gdpr/common tests pass.

The email-wrapper consolidation (audit rec #6) is intentionally left out of this
behavior-preserving split — it changes task signatures and belongs in its own
commit with focused tests.
