--- src/accounts/tasks.py
@@ -1,995 +1,0 @@
-"""Tasks for the authentication app."""
-
-import traceback
-import typing as t
-from datetime import datetime, timedelta
-
-import httpx
-import stripe
-import structlog
-from celery import shared_task
-from django.conf import settings
-from django.contrib.sites.models import Site
-from django.db import transaction
-from django.db.models import Q
-from django.template.loader import render_to_string
-from django.utils import timezone
-from django.utils.translation import gettext as _
-from ninja_jwt.token_blacklist.models import OutstandingToken
-from ninja_jwt.utils import aware_utcnow
-
-from accounts.models import (
-    EmailVerificationReminderTracking,
-    GlobalBan,
-    ReferralPayout,
-    ReferralPayoutStatement,
-    RevelUser,
-    UserDataExport,
-)
-from accounts.service import gdpr
-from common.models import SiteSettings
-from common.signing import generate_signed_url
-from common.tasks import send_email
-from events.utils.currency import to_stripe_amount
-
-# 7 days in seconds for data export download links
-DATA_EXPORT_URL_EXPIRES_IN = 7 * 24 * 60 * 60
-
-logger = structlog.get_logger(__name__)
-
-
-@shared_task
-def send_verification_email(email: str, token: str) -> None:
-    """Send a verification email."""
-    logger.info("verification_email_sending", email=email)
-    subject = str(render_to_string("accounts/emails/email_verification_subject.txt"))
-    verification_link = SiteSettings.get_solo().frontend_base_url + f"/login/confirm-email?token={token}"
-    body = render_to_string("accounts/emails/email_verification_body.txt", {"verification_link": verification_link})
-    html_body = render_to_string(
-        "accounts/emails/email_verification_body.html", {"verification_link": verification_link}
-    )
-    send_email(to=email, subject=subject, body=body, html_body=html_body)
-    logger.info("verification_email_sent", email=email)
-
-
-@shared_task
-def send_account_activation_link(email: str, token: str) -> None:
-    """Send an account activation email to a guest user registering for a full account."""
-    logger.info("account_activation_email_sending", email=email)
-    subject = str(render_to_string("accounts/emails/account_activation_subject.txt"))
-    activation_link = SiteSettings.get_solo().frontend_base_url + f"/login/reset-password?token={token}"
-    body = render_to_string("accounts/emails/account_activation_body.txt", {"activation_link": activation_link})
-    html_body = render_to_string("accounts/emails/account_activation_body.html", {"activation_link": activation_link})
-    send_email(to=email, subject=subject, body=body, html_body=html_body)
-    logger.info("account_activation_email_sent", email=email)
-
-
-@shared_task
-def send_password_reset_link(email: str, token: str) -> None:
-    """Send a password reset email."""
-    logger.info("password_reset_email_sending", email=email)
-    subject = str(render_to_string("accounts/emails/password_reset_subject.txt"))
-    password_reset_link = SiteSettings.get_solo().frontend_base_url + f"/login/reset-password?token={token}"
-    body = render_to_string("accounts/emails/password_reset_body.txt", {"password_reset_link": password_reset_link})
-    html_body = render_to_string(
-        "accounts/emails/password_reset_body.html", {"password_reset_link": password_reset_link}
-    )
-    send_email(to=email, subject=subject, body=body, html_body=html_body)
-    logger.info("password_reset_email_sent", email=email)
-
-
-@shared_task
-def send_email_change_confirmation(new_email: str, token: str) -> None:
-    """Send the confirmation link to the **new** email address.
-
-    Clicking the link proves the user controls the new mailbox.
-
-    Args:
-        new_email: The new email address requested by the user.
-        token: The single-use email-change JWT to embed in the confirmation link.
-    """
-    logger.info("email_change_confirmation_sending", new_email=new_email)
-    subject = str(render_to_string("accounts/emails/email_change_confirmation_subject.txt"))
-    site_settings = SiteSettings.get_solo()
-    confirmation_link = site_settings.frontend_base_url + f"/account/confirm-email-change?token={token}"
-    context = {"confirmation_link": confirmation_link, "frontend_base_url": site_settings.frontend_base_url}
-    body = render_to_string("accounts/emails/email_change_confirmation_body.txt", context)
-    html_body = render_to_string("accounts/emails/email_change_confirmation_body.html", context)
-    send_email(to=new_email, subject=subject, body=body, html_body=html_body)
-    logger.info("email_change_confirmation_sent", new_email=new_email)
-
-
-@shared_task
-def send_email_change_notice(current_email: str, masked_new_email: str) -> None:
-    """Notify the **current** email address that a change was requested.
-
-    Informational only — there is no cancel link in v1.
-
-    Args:
-        current_email: The user's current email address.
-        masked_new_email: The new address in masked form (e.g. ``a***@example.com``).
-    """
-    logger.info("email_change_notice_sending", current_email=current_email)
-    subject = str(render_to_string("accounts/emails/email_change_notice_subject.txt"))
-    site_settings = SiteSettings.get_solo()
-    context = {"masked_new_email": masked_new_email, "frontend_base_url": site_settings.frontend_base_url}
-    body = render_to_string("accounts/emails/email_change_notice_body.txt", context)
-    html_body = render_to_string("accounts/emails/email_change_notice_body.html", context)
-    send_email(to=current_email, subject=subject, body=body, html_body=html_body)
-    logger.info("email_change_notice_sent", current_email=current_email)
-
-
-@shared_task
-def send_email_change_completed_old(old_email: str, new_email: str) -> None:
-    """Notify the **old** address that the email change has completed.
-
-    Args:
-        old_email: The address being decommissioned.
-        new_email: The address that is now primary on the account.
-    """
-    logger.info("email_change_completed_old_sending", old_email=old_email)
-    subject = str(render_to_string("accounts/emails/email_change_completed_old_subject.txt"))
-    site_settings = SiteSettings.get_solo()
-    context = {"old_email": old_email, "new_email": new_email, "frontend_base_url": site_settings.frontend_base_url}
-    body = render_to_string("accounts/emails/email_change_completed_old_body.txt", context)
-    html_body = render_to_string("accounts/emails/email_change_completed_old_body.html", context)
-    send_email(to=old_email, subject=subject, body=body, html_body=html_body)
-    logger.info("email_change_completed_old_sent", old_email=old_email)
-
-
-@shared_task
-def send_email_change_completed_new(new_email: str, old_email: str) -> None:
-    """Welcome message to the **new** address confirming the change is live.
-
-    Args:
-        new_email: The address that is now primary on the account.
-        old_email: The previous address (referenced in the body for clarity).
-    """
-    logger.info("email_change_completed_new_sending", new_email=new_email)
-    subject = str(render_to_string("accounts/emails/email_change_completed_new_subject.txt"))
-    site_settings = SiteSettings.get_solo()
-    context = {"old_email": old_email, "new_email": new_email, "frontend_base_url": site_settings.frontend_base_url}
-    body = render_to_string("accounts/emails/email_change_completed_new_body.txt", context)
-    html_body = render_to_string("accounts/emails/email_change_completed_new_body.html", context)
-    send_email(to=new_email, subject=subject, body=body, html_body=html_body)
-    logger.info("email_change_completed_new_sent", new_email=new_email)
-
-
-@shared_task
-def send_account_deletion_link(email: str, token: str) -> None:
-    """Send an account deletion confirmation email."""
-    logger.info("account_deletion_email_sending", email=email)
-    site = Site.objects.get_current()
-    subject = str(render_to_string("accounts/emails/account_delete_subject.txt", {"site_name": site.name}))
-    site_settings = SiteSettings.get_solo()
-    account_deletion_link = site_settings.frontend_base_url + f"/account/confirm-deletion?token={token}"
-    body = render_to_string(
-        "accounts/emails/account_delete_body.txt",
-        {"account_deletion_link": account_deletion_link, "frontend_base_url": site_settings.frontend_base_url},
-    )
-    html_body = render_to_string(
-        "accounts/emails/account_delete_body.html",
-        {"account_deletion_link": account_deletion_link, "frontend_base_url": site_settings.frontend_base_url},
-    )
-    send_email(to=email, subject=subject, body=body, html_body=html_body)
-    logger.info("account_deletion_email_sent", email=email)
-
-
-@shared_task
-def flush_expired_tokens() -> None:
-    """Flushes any expired tokens in the outstanding token list.
-
-    This task is designed to be run periodically to clean up expired tokens.
-    """
-    logger.info("token_cleanup_started")
-    # Get the current time in UTC
-    current_time = aware_utcnow()
-
-    # Delete expired tokens
-    jwt_deleted, _ = OutstandingToken.objects.filter(expires_at__lte=current_time).delete()
-    logger.info("token_cleanup_completed", jwt_tokens_deleted=jwt_deleted)
-
-
-@shared_task
-def generate_user_data_export(user_id: str) -> None:
-    """Generate a data export for a user."""
-    logger.info("gdpr_export_task_started", user_id=user_id)
-    user = RevelUser.objects.get(id=user_id)
-    try:
-        data_export = gdpr.generate_user_data_export(user)
-    except Exception as e:
-        logger.error("gdpr_export_task_failed", user_id=user_id, error=str(e), exc_info=True)
-        _notify_data_export_failed(user, traceback.format_exc())
-        raise  # Re-raise so Celery marks task as failed for monitoring/alerting
-    logger.info("gdpr_export_task_completed", user_id=user_id, export_id=str(data_export.id))
-    _notify_user_data_export_ready(data_export)
-
-
-def _notify_data_export_failed(user: RevelUser, error: str) -> None:
-    logger.info("gdpr_export_notification_failed", user_id=str(user.id), email=user.email)
-    data_export, _ = UserDataExport.objects.get_or_create(user=user)
-    data_export.status = UserDataExport.UserDataExportStatus.FAILED
-    data_export.error_message = error
-    data_export.save(update_fields=["status", "error_message"])
-
-    # Notify the user that something went wrong
-    subject = str(render_to_string("accounts/emails/data_export_failed_subject.txt"))
-    body = render_to_string("accounts/emails/data_export_failed_body.txt", {"user": user})
-    html_body = render_to_string("accounts/emails/data_export_failed_body.html", {"user": user})
-    send_email(to=user.email, subject=subject, body=body, html_body=html_body)
-
-    # Notify admins
-    subject = str(render_to_string("accounts/emails/data_export_failed_admin_subject.txt"))
-    admins = RevelUser.objects.filter(Q(is_superuser=True) | Q(is_staff=True))
-    admin_count = admins.count()
-    logger.info("gdpr_export_admin_notification_sending", user_id=str(user.id), admin_count=admin_count)
-    for admin in admins:
-        body = render_to_string(
-            "accounts/emails/data_export_failed_admin_body.txt",
-            {"user": user, "error_message": data_export.error_message},
-        )
-        html_body = render_to_string(
-            "accounts/emails/data_export_failed_admin_body.html",
-            {"user": user, "error_message": data_export.error_message},
-        )
-        send_email(to=admin.email, subject=subject, body=body, html_body=html_body)
-
-
-def _notify_user_data_export_ready(data_export: UserDataExport) -> None:
-    logger.info(
-        "gdpr_export_notification_ready",
-        user_id=str(data_export.user.id),
-        email=data_export.user.email,
-        export_id=str(data_export.id),
-    )
-    signed_path = generate_signed_url(data_export.file.name, expires_in=DATA_EXPORT_URL_EXPIRES_IN)
-    download_url = settings.BASE_URL + signed_path
-    subject = "Your Revel Data Export is Ready"
-    body = render_to_string(
-        "accounts/emails/data_export_ready_body.txt", {"download_url": download_url, "user": data_export.user}
-    )
-    html_body = render_to_string(
-        "accounts/emails/data_export_ready_body.html", {"download_url": download_url, "user": data_export.user}
-    )
-    send_email(to=data_export.user.email, subject=subject, body=body, html_body=html_body)
-
-
-@shared_task
-def delete_user_account(user_id: str) -> None:
-    """Delete a user account and all associated data in the background.
-
-    This task is designed to handle heavy deletion operations that may involve
-    many database relationships. The deletion is performed in a transaction
-    to ensure data consistency.
-
-    Args:
-        user_id: The UUID of the user to delete.
-    """
-    user = RevelUser.objects.get(id=user_id)
-    logger.info("account_deletion_started", user_id=str(user.id), email=user.email)
-    try:
-        user.delete()
-        logger.info("account_deletion_completed", user_id=user_id)
-    except Exception as e:
-        logger.error("account_deletion_failed", user_id=user_id, error=str(e), exc_info=True)
-        raise
-
-
-def _get_referrer_email(user_id: str) -> str | None:
-    """Return the email of the user who referred `user_id`, or None if none."""
-    from uuid import UUID
-
-    from accounts.models import Referral
-
-    referral = Referral.objects.select_related("referrer").filter(referred_user_id=UUID(user_id)).first()
-    return referral.referrer.email if referral else None
-
-
-@shared_task(bind=True, max_retries=3)
-def notify_admin_new_user_joined(self: t.Any, user_id: str, user_email: str, is_guest: bool) -> dict[str, t.Any]:
-    """Send Pushover notification to admin when a new user joins the platform.
-
-    This is a standalone notification system separate from the main notification infrastructure.
-    No database records are created - this simply sends a Pushover notification if configured.
-
-    Args:
-        self: Celery task instance (automatically passed when bind=True)
-        user_id: UUID of the user who joined
-        user_email: Email of the user who joined
-        is_guest: Whether the user is a guest user
-
-    Returns:
-        Dict with notification result
-
-    Raises:
-        Exception: If Pushover API call fails after retries
-    """
-    if not settings.PUSHOVER_USER_KEY or not settings.PUSHOVER_APP_TOKEN:
-        logger.warning(
-            "pushover_not_configured",
-            message="PUSHOVER_USER_KEY or PUSHOVER_APP_TOKEN not set in settings",
-        )
-        return {"status": "skipped", "reason": "pushover_not_configured"}
-
-    user_type_label = "guest user" if is_guest else "user"
-    title = "New Guest User" if is_guest else "New User"
-    referrer_email = _get_referrer_email(user_id)
-    user_count = RevelUser.objects.filter(guest=False).count()
-
-    lines = [f"New {user_type_label} joined: {user_email}"]
-    if referrer_email:
-        lines.append(f"Referred by: {referrer_email}")
-    lines.append(f"We now have {user_count} users!")
-    message = "\n".join(lines)
-
-    pushover_url = "https://api.pushover.net/1/messages.json"
-    payload = {
-        "token": settings.PUSHOVER_APP_TOKEN,
-        "user": settings.PUSHOVER_USER_KEY,
-        "message": message,
-        "title": title,
-        "priority": 0,
-    }
-
-    try:
-        response = httpx.post(pushover_url, data=payload, timeout=10.0)
-        response.raise_for_status()
-
-        logger.info(
-            "pushover_notification_sent",
-            user_id=user_id,
-            user_email=user_email,
-            is_guest=is_guest,
-            response_status=response.status_code,
-        )
-
-        return {
-            "status": "sent",
-            "user_id": user_id,
-            "user_email": user_email,
-            "is_guest": is_guest,
-        }
-
-    except (httpx.HTTPStatusError, httpx.RequestError, httpx.TimeoutException) as e:
-        logger.error(
-            "pushover_error",
-            user_id=user_id,
-            error=str(e),
-        )
-        if self.request.retries < self.max_retries:
-            countdown = 2**self.request.retries * 60  # 1min, 2min, 4min
-            logger.info(
-                "retrying_pushover_notification",
-                user_id=user_id,
-                countdown=countdown,
-                retry_count=self.request.retries,
-            )
-            raise self.retry(exc=e, countdown=countdown)
-        else:
-            logger.exception(
-                "pushover_exception",
-                user_id=user_id,
-                error=str(e),
-            )
-            raise
-
-
-@shared_task(bind=True, max_retries=3)
-def notify_admin_new_user_joined_discord(self: t.Any) -> dict[str, t.Any]:
-    """Send a PII-free Discord notification to admin when a new user joins.
-
-    The message never contains the user's email, name, or id — only the running
-    total count of non-guest users.
-
-    Returns:
-        Dict with notification result
-
-    Raises:
-        Exception: If the Discord webhook call fails after retries.
-    """
-    webhook_url = settings.DISCORD_ADMIN_WEBHOOK_URL
-    if not webhook_url:
-        logger.info("discord_webhook_not_configured")
-        return {"status": "skipped", "reason": "discord_webhook_not_configured"}
-
-    user_count = RevelUser.objects.filter(guest=False).count()
-    payload = {
-        "content": f"🎉 A new user joined! We now have {user_count} users.",
-        "allowed_mentions": {"parse": []},
-    }
-
-    try:
-        response = httpx.post(webhook_url, json=payload, timeout=10.0)
-        response.raise_for_status()
-        logger.info(
-            "discord_notification_sent",
-            channel="user_joined",
-            user_count=user_count,
-            response_status=response.status_code,
-        )
-        return {"status": "sent", "user_count": user_count}
-    except (httpx.HTTPStatusError, httpx.RequestError, httpx.TimeoutException) as e:
-        logger.error("discord_error", channel="user_joined", error=str(e))
-        if self.request.retries < self.max_retries:
-            countdown = 2**self.request.retries * 60
-            raise self.retry(exc=e, countdown=countdown)
-        logger.exception("discord_exception", channel="user_joined", error=str(e))
-        raise
-
-
-@shared_task(bind=True, max_retries=3)
-def process_domain_ban_task(self: t.Any, ban_id: str) -> dict[str, t.Any]:
-    """Process a domain ban asynchronously, deactivating all matching users.
-
-    The task is idempotent: already-deactivated users are skipped on retry.
-
-    Args:
-        self: Celery task instance (automatically passed when bind=True).
-        ban_id: UUID of the GlobalBan instance.
-
-    Returns:
-        Dict with domain and deactivated_count.
-    """
-    from accounts.service.global_ban_service import process_domain_ban
-
-    ban = GlobalBan.objects.get(id=ban_id)
-
-    try:
-        count = process_domain_ban(ban)
-        logger.info("domain_ban_task_completed", ban_id=ban_id, domain=ban.value, deactivated_count=count)
-        return {"domain": ban.value, "deactivated_count": count}
-    except Exception as exc:
-        logger.error("domain_ban_task_failed", ban_id=ban_id, error=str(exc), retry=self.request.retries)
-        raise self.retry(exc=exc, countdown=2**self.request.retries * 60)
-
-
-# Email verification reminder tasks
-
-
-def mark_reminder_sent(*, user_id: str, reminder_type: str, success: bool, error_message: str | None = None) -> None:
-    """Callback function to update tracking after reminder email send attempt.
-
-    Args:
-        user_id: UUID of the user.
-        reminder_type: Either "early", "final_warning", or "deactivation".
-        success: Whether the email was sent successfully.
-        error_message: Error message if sending failed.
-    """
-    if not success:
-        logger.warning(
-            "reminder_email_failed",
-            user_id=user_id,
-            reminder_type=reminder_type,
-            error=error_message,
-        )
-        return
-
-    try:
-        user = RevelUser.objects.get(id=user_id)
-        tracking, _ = EmailVerificationReminderTracking.objects.get_or_create(user=user)
-        now = timezone.now()
-
-        if reminder_type == "early":
-            tracking.last_reminder_sent_at = now
-            tracking.save(update_fields=["last_reminder_sent_at"])
-        elif reminder_type == "final_warning":
-            tracking.final_warning_sent_at = now
-            tracking.last_reminder_sent_at = now
-            tracking.save(update_fields=["final_warning_sent_at", "last_reminder_sent_at"])
-        elif reminder_type == "deactivation":
-            tracking.deactivation_email_sent_at = now
-            tracking.save(update_fields=["deactivation_email_sent_at"])
-        else:
-            logger.error(
-                "invalid_reminder_type",
-                user_id=user_id,
-                reminder_type=reminder_type,
-            )
-            return
-
-        logger.info(
-            "reminder_tracking_updated",
-            user_id=user_id,
-            reminder_type=reminder_type,
-        )
-    except RevelUser.DoesNotExist:
-        logger.warning("reminder_callback_user_not_found", user_id=user_id)
-
-
-def _send_reminder_email(user: RevelUser, reminder_type: str) -> None:
-    """Send verification reminder email with both verification and deletion links.
-
-    Args:
-        user: The user to send reminder to.
-        reminder_type: Either "early" or "final_warning".
-    """
-    from accounts.service.account import create_deletion_token, create_verification_token
-
-    verification_token = create_verification_token(user)
-    deletion_token = create_deletion_token(user)
-
-    site_settings = SiteSettings.get_solo()
-    verification_link = site_settings.frontend_base_url + f"/login/confirm-email?token={verification_token}"
-    deletion_link = site_settings.frontend_base_url + f"/account/confirm-deletion?token={deletion_token}"
-
-    if reminder_type == "final_warning":
-        subject = render_to_string("accounts/emails/email_verification_final_warning_subject.txt").strip()
-        template_base = "accounts/emails/email_verification_final_warning"
-    else:
-        subject = render_to_string("accounts/emails/email_verification_reminder_subject.txt").strip()
-        template_base = "accounts/emails/email_verification_reminder"
-
-    context = {
-        "verification_link": verification_link,
-        "deletion_link": deletion_link,
-        "user": user,
-    }
-
-    body = render_to_string(f"{template_base}_body.txt", context)
-    html_body = render_to_string(f"{template_base}_body.html", context)
-
-    callback_data = {
-        "module": "accounts.tasks",
-        "function": "mark_reminder_sent",
-        "kwargs": {
-            "user_id": str(user.id),
-            "reminder_type": reminder_type,
-        },
-    }
-
-    send_email.delay(to=user.email, subject=subject, body=body, html_body=html_body, callback_data=callback_data)
-    logger.info("verification_reminder_queued", user_id=str(user.id), email=user.email, reminder_type=reminder_type)
-
-
-def _send_deactivation_email(user: RevelUser) -> None:
-    """Send account deactivation notice with verification and deletion links.
-
-    Args:
-        user: The deactivated user.
-    """
-    from accounts.service.account import create_deletion_token, create_verification_token
-
-    verification_token = create_verification_token(user)
-    deletion_token = create_deletion_token(user)
-
-    site_settings = SiteSettings.get_solo()
-    verification_link = site_settings.frontend_base_url + f"/login/confirm-email?token={verification_token}"
-    deletion_link = site_settings.frontend_base_url + f"/account/confirm-deletion?token={deletion_token}"
-
-    subject = render_to_string("accounts/emails/account_deactivated_subject.txt").strip()
-    context = {
-        "verification_link": verification_link,
-        "deletion_link": deletion_link,
-        "user": user,
-    }
-
-    body = render_to_string("accounts/emails/account_deactivated_body.txt", context)
-    html_body = render_to_string("accounts/emails/account_deactivated_body.html", context)
-
-    callback_data = {
-        "module": "accounts.tasks",
-        "function": "mark_reminder_sent",
-        "kwargs": {
-            "user_id": str(user.id),
-            "reminder_type": "deactivation",
-        },
-    }
-
-    send_email.delay(to=user.email, subject=subject, body=body, html_body=html_body, callback_data=callback_data)
-    logger.info("deactivation_email_queued", user_id=str(user.id), email=user.email)
-
-
-def _process_reminder_batch(users: t.Any, reminder_age_cutoff: datetime) -> int:
-    """Process a batch of users for reminder sending.
-
-    Args:
-        users: QuerySet of users to process.
-        reminder_age_cutoff: Only send if last reminder was before this time.
-
-    Returns:
-        Count of reminders sent.
-    """
-    count = 0
-    for user in users:
-        tracking = getattr(user, "verification_reminder_tracking", None)
-        if tracking is None:
-            tracking = EmailVerificationReminderTracking.objects.create(user=user)
-            user.verification_reminder_tracking = tracking  # Cache for subsequent loops
-        # Skip if final warning already sent
-        if tracking.final_warning_sent_at:
-            continue
-        # Send if never sent OR sent before cutoff
-        if not tracking.last_reminder_sent_at or tracking.last_reminder_sent_at <= reminder_age_cutoff:
-            _send_reminder_email(user, "early")
-            count += 1
-    return count
-
-
-@shared_task
-def send_early_verification_reminders() -> dict[str, int]:
-    """Send verification reminders for accounts 24h-30d old.
-
-    Backoff schedule:
-    - 24h+ old: every 24 hours
-    - 3d+ old: every 48 hours
-    - 7d+ old: every 7 days
-
-    Returns:
-        Dict with counts of reminders sent per stage.
-    """
-    now = timezone.now()
-    stats = {"24h": 0, "3d": 0, "7d": 0}
-
-    # Base queryset: unverified, active, non-guest users with tracking prefetched
-    base_qs = RevelUser.objects.filter(email_verified=False, is_active=True, guest=False).select_related(
-        "verification_reminder_tracking"
-    )
-
-    # 24-hour reminders (account 24h-3d old, last reminder >24h ago OR never sent)
-    day_ago = now - timedelta(hours=24)
-    three_days_ago = now - timedelta(days=3)
-    users_24h = base_qs.filter(date_joined__lte=day_ago, date_joined__gt=three_days_ago)
-    stats["24h"] = _process_reminder_batch(users_24h, day_ago)
-
-    # 3-day reminders (account 3d-7d old, last reminder >48h ago)
-    two_days_ago = now - timedelta(hours=48)
-    seven_days_ago = now - timedelta(days=7)
-    users_3d = base_qs.filter(date_joined__lte=three_days_ago, date_joined__gt=seven_days_ago)
-    stats["3d"] = _process_reminder_batch(users_3d, two_days_ago)
-
-    # 7-day reminders (account 7d-30d old, last reminder >7d ago)
-    week_ago = now - timedelta(days=7)
-    month_ago = now - timedelta(days=30)
-    users_7d = base_qs.filter(date_joined__lte=seven_days_ago, date_joined__gt=month_ago)
-    stats["7d"] = _process_reminder_batch(users_7d, week_ago)
-
-    logger.info("early_verification_reminders_sent", **stats)
-    return stats
-
-
-@shared_task
-def send_final_verification_warnings() -> dict[str, int]:
-    """Send final warning to accounts 30+ days old that never verified.
-
-    This warning is sent ONCE and informs users their account will be
-    deactivated shortly if they don't verify.
-
-    Returns:
-        Dict with count of final warnings sent.
-    """
-    now = timezone.now()
-    month_ago = now - timedelta(days=30)
-
-    users_needing_final_warning = RevelUser.objects.filter(
-        email_verified=False, is_active=True, guest=False, date_joined__lte=month_ago
-    ).select_related("verification_reminder_tracking")
-
-    count = 0
-    for user in users_needing_final_warning:
-        tracking = getattr(user, "verification_reminder_tracking", None)
-        if tracking is None:
-            tracking = EmailVerificationReminderTracking.objects.create(user=user)
-            user.verification_reminder_tracking = tracking  # Cache the created tracking
-        # Only send if final warning hasn't been sent yet
-        if not tracking.final_warning_sent_at:
-            _send_reminder_email(user, "final_warning")
-            count += 1
-            logger.info("final_verification_warning_queued", user_id=str(user.id), email=user.email)
-
-    logger.info("final_verification_warnings_sent", count=count)
-    return {"count": count}
-
-
-@shared_task
-def deactivate_unverified_accounts() -> dict[str, int]:
-    """Deactivate accounts that received final warning and still haven't verified.
-
-    Accounts are deactivated at least 24 hours after receiving the final warning
-    if they still haven't verified. They receive one more email with verification
-    link and deletion link.
-
-    Returns:
-        Dict with count of accounts deactivated.
-    """
-    now = timezone.now()
-    grace_period = timedelta(hours=24)
-    grace_cutoff = now - grace_period
-
-    # Get all tracking records with final warning sent >24h ago but no deactivation email
-    tracking_records = EmailVerificationReminderTracking.objects.filter(
-        final_warning_sent_at__lte=grace_cutoff, deactivation_email_sent_at__isnull=True
-    ).select_related("user")
-
-    count = 0
-    for tracking in tracking_records:
-        user = tracking.user
-        # Double-check user is still unverified and active
-        if not user.email_verified and user.is_active:
-            user.is_active = False
-            user.save(update_fields=["is_active"])
-
-            _send_deactivation_email(user)
-            count += 1
-            logger.warning(
-                "account_deactivated_unverified",
-                user_id=str(user.id),
-                email=user.email,
-                account_age_days=(now - user.date_joined).days,
-            )
-
-    logger.info("accounts_deactivated", count=count)
-    return {"count": count}
-
-
-@shared_task
-def cleanup_expired_data_exports() -> dict[str, int]:
-    """Delete files from expired data exports while preserving database records.
-
-    Data export download links are valid for 7 days. After that, the file is no longer
-    accessible via signed URL, so we can safely delete it to reclaim storage.
-
-    The database record is preserved for auditing purposes (shows user requested an export).
-
-    Returns:
-        Dict with count of files deleted.
-    """
-    now = timezone.now()
-    expiry_cutoff = now - timedelta(seconds=DATA_EXPORT_URL_EXPIRES_IN)
-
-    # Find exports with files that are older than the URL expiry time
-    expired_exports = UserDataExport.objects.filter(
-        completed_at__lte=expiry_cutoff,
-        status=UserDataExport.UserDataExportStatus.READY,
-    ).exclude(file="")
-
-    count = 0
-    for export in expired_exports:
-        export.file.delete(save=False)
-        export.file = None
-        export.save(update_fields=["file", "updated_at"])
-        count += 1
-        logger.info(
-            "data_export_file_deleted",
-            export_id=str(export.id),
-            user_id=str(export.user_id),
-            completed_at=export.completed_at.isoformat() if export.completed_at else None,
-        )
-
-    logger.info("data_export_cleanup_completed", files_deleted=count)
-    return {"files_deleted": count}
-
-
-@shared_task
-def delete_old_inactive_accounts() -> dict[str, t.Any]:
-    """Permanently delete accounts inactive for 60+ days.
-
-    Accounts are deleted if they were deactivated (deactivation_email_sent_at)
-    60 or more days ago.
-
-    Returns:
-        Dict with count of accounts deleted and list of deleted user info.
-    """
-    now = timezone.now()
-    sixty_days_ago = now - timedelta(days=60)
-
-    # Get all tracking records with deactivation email sent 60+ days ago
-    tracking_records = EmailVerificationReminderTracking.objects.filter(
-        deactivation_email_sent_at__lte=sixty_days_ago
-    ).select_related("user")
-
-    deleted_users = []
-    for tracking in tracking_records:
-        user = tracking.user
-        # Double-check user is still inactive and unverified
-        if not user.is_active and not user.email_verified:
-            user_info = {
-                "user_id": str(user.id),
-                "email": user.email,
-                "date_joined": user.date_joined.isoformat(),
-                "deactivated_at": tracking.deactivation_email_sent_at.isoformat()
-                if tracking.deactivation_email_sent_at
-                else None,
-            }
-            deleted_users.append(user_info)
-            logger.warning("deleting_old_inactive_account", **user_info)
-            user.delete()  # This will cascade delete the tracking record too
-
-    logger.info("old_inactive_accounts_deleted", count=len(deleted_users))
-    return {"count": len(deleted_users), "deleted_users": deleted_users}
-
-
-# ---------------------------------------------------------------------------
-# Referral payout processing
-# ---------------------------------------------------------------------------
-
-stripe.api_key = settings.STRIPE_SECRET_KEY
-stripe.api_version = settings.STRIPE_API_VERSION
-
-
-def _reclaim_stale_payouts() -> None:
-    """Reclaim PENDING payouts abandoned by crashed workers (stale > 1 hour)."""
-    stale_cutoff = timezone.now() - timedelta(hours=1)
-    stale_count = ReferralPayout.objects.filter(
-        status=ReferralPayout.ReferralPayoutStatus.PENDING,
-        updated_at__lt=stale_cutoff,
-    ).update(status=ReferralPayout.ReferralPayoutStatus.CALCULATED)
-    if stale_count:
-        logger.warning("stale_pending_payouts_reclaimed", count=stale_count)
-
-
-def _validate_payout_eligibility(payout: ReferralPayout, referrer: RevelUser) -> str | None:
-    """Run pre-flight checks for a payout. Returns a skip reason or None if eligible."""
-    if not referrer.stripe_account_id or not referrer.stripe_charges_enabled:
-        return "payout_skipped_no_stripe"
-
-    if not hasattr(referrer, "billing_profile"):
-        return "payout_skipped_no_billing"
-
-    profile = referrer.billing_profile
-    if not profile.self_billing_agreed:
-        return "payout_skipped_no_agreement"
-
-    missing_fields = [f for f in ("billing_name", "vat_country_code", "billing_address") if not getattr(profile, f)]
-    if missing_fields:
-        logger.info(
-            "payout_skipped_incomplete_billing",
-            referrer_id=str(referrer.id),
-            payout_id=str(payout.id),
-            missing=missing_fields,
-        )
-        return "payout_skipped_incomplete_billing"
-
-    if payout.payout_amount < settings.MINIMUM_PAYOUT_AMOUNT:
-        logger.info(
-            "payout_skipped_below_threshold",
-            referrer_id=str(referrer.id),
-            payout_id=str(payout.id),
-            amount=str(payout.payout_amount),
-            threshold=str(settings.MINIMUM_PAYOUT_AMOUNT),
-        )
-        return "payout_skipped_below_threshold"
-
-    return None
-
-
-@shared_task(name="accounts.process_referral_payouts")
-def process_referral_payouts() -> dict[str, int]:
-    """Process all CALCULATED referral payouts via Stripe Transfer.
-
-    For each payout:
-    1. Claim the row (``CALCULATED → PENDING`` under ``select_for_update``).
-    2. Verify referrer has a connected Stripe account, a billing profile,
-       and has agreed to self-billing.
-    3. Create a Stripe Transfer (with idempotency key ``payout.id``).
-    4. On success → ``PAID``; on Stripe error → ``FAILED`` (logged, loop continues).
-    5. Generate the payout statement PDF and email it to the referrer.
-
-    Errors are handled per-payout so one failure does not block the rest.
-
-    Returns:
-        Dict with ``paid``, ``failed``, and ``skipped`` counts.
-    """
-    from accounts.service.payout_statement_service import generate_payout_statement
-
-    _reclaim_stale_payouts()
-
-    payout_ids = list(
-        ReferralPayout.objects.filter(status=ReferralPayout.ReferralPayoutStatus.CALCULATED)
-        .order_by("period_start")
-        .values_list("id", flat=True)
-    )
-
-    stats: dict[str, int] = {"paid": 0, "failed": 0, "skipped": 0}
-
-    for payout_id in payout_ids:
-        # Claim the payout under a row lock to prevent duplicate processing.
-        with transaction.atomic():
-            payout = (
-                ReferralPayout.objects.select_for_update(skip_locked=True)
-                .filter(id=payout_id, status=ReferralPayout.ReferralPayoutStatus.CALCULATED)
-                .first()
-            )
-            if payout is None:
-                continue
-            payout.status = ReferralPayout.ReferralPayoutStatus.PENDING
-            payout.save(update_fields=["status", "updated_at"])
-
-        # Reload with relations outside the lock
-        payout = ReferralPayout.objects.select_related("referral__referrer__billing_profile", "referral__referrer").get(
-            id=payout_id
-        )
-        referrer = payout.referral.referrer
-
-        skip_reason = _validate_payout_eligibility(payout, referrer)
-        if skip_reason:
-            if skip_reason not in ("payout_skipped_incomplete_billing", "payout_skipped_below_threshold"):
-                logger.info(skip_reason, referrer_id=str(referrer.id), payout_id=str(payout.id))
-            payout.status = ReferralPayout.ReferralPayoutStatus.CALCULATED
-            payout.save(update_fields=["status", "updated_at"])
-            stats["skipped"] += 1
-            continue
-
-        # --- Stripe Transfer (idempotent via payout ID key) ---
-        try:
-            transfer = stripe.Transfer.create(
-                amount=to_stripe_amount(payout.payout_amount, payout.currency),
-                currency=payout.currency.lower(),
-                destination=referrer.stripe_account_id,
-                transfer_group=f"referral-payout-{payout.id}",
-                idempotency_key=f"referral-payout-{payout.id}",
-            )
-        except stripe.error.StripeError as exc:
-            payout.status = ReferralPayout.ReferralPayoutStatus.FAILED
-            payout.save(update_fields=["status", "updated_at"])
-            stats["failed"] += 1
-
-            logger.error(
-                "payout_transfer_failed",
-                payout_id=str(payout.id),
-                referrer_id=str(referrer.id),
-                error=str(exc),
-            )
-            continue
-
-        payout.status = ReferralPayout.ReferralPayoutStatus.PAID
-        payout.stripe_transfer_id = transfer.id
-        payout.save(update_fields=["status", "stripe_transfer_id", "updated_at"])
-        stats["paid"] += 1
-
-        logger.info(
-            "payout_transferred",
-            payout_id=str(payout.id),
-            transfer_id=transfer.id,
-            amount=str(payout.payout_amount),
-            currency=payout.currency,
-        )
-
-        # --- Generate statement + email (only after successful transfer) ---
-        statement = generate_payout_statement(payout)
-        _send_payout_statement_email(payout, statement, referrer)
-
-    logger.info("process_referral_payouts_completed", **stats)
-    return stats
-
-
-def _send_payout_statement_email(
-    payout: ReferralPayout,
-    statement: ReferralPayoutStatement,
-    referrer: RevelUser,
-) -> None:
-    """Dispatch the payout statement PDF to the referrer via email."""
-    billing_profile = getattr(referrer, "billing_profile", None)
-    recipient = billing_profile.billing_email if billing_profile and billing_profile.billing_email else referrer.email
-
-    subject = _("Referral payout statement %(document_number)s (%(currency)s)") % {
-        "document_number": statement.document_number,
-        "currency": payout.currency,
-    }
-    body = _(
-        "Please find attached your referral payout statement %(document_number)s "
-        "for the period %(period_start)s to %(period_end)s."
-    ) % {
-        "document_number": statement.document_number,
-        "period_start": payout.period_start.isoformat(),
-        "period_end": payout.period_end.isoformat(),
-    }
-
-    site = SiteSettings.get_solo()
-    bcc = [site.platform_invoice_bcc_email] if site.platform_invoice_bcc_email else []
-
-    send_email(
-        to=recipient,
-        subject=subject,
-        body=body,
-        bcc=bcc,
-        from_email=settings.DEFAULT_BILLING_EMAIL,
-        reply_to=[settings.DEFAULT_REPLY_TO_EMAIL],
-        attachment_storage_path=statement.pdf_file.name,
-        attachment_filename=f"{statement.document_number}.pdf",
-    )
-
-    logger.info(
-        "payout_statement_email_sent",
-        document_number=statement.document_number,
-        referrer_id=str(referrer.id),
-    )
